requirements.txt
# Core Libraries
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.1
numpy==1.24.2
pandas==1.5.3
matplotlib==3.7.1
scikit-learn==1.2.2
opencv-python==4.7.0.72
Pillow==9.5.0
albumentations==1.3.0

# Deep Learning Models
ultralytics==8.0.0  # For YOLOv8

# Web Application
streamlit==1.25.0

# Utilities
tqdm==4.65.0
